some staff

+++ src/main/java/emsbj/SchoolYearRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface SchoolYearRepository extends CrudRepository<SchoolYear, Integer> {
+public interface SchoolYearRepository extends CrudRepository<SchoolYear, Long> {
     Optional<SchoolYear> findByBeginYear(Integer beginYear);
 }
